refactor: clean up assets/icons into subfolders and change color
